Allow RPC framework to use rank in addition to WorkerInfo and name.
The RPC framework only allowed sending RPCs based on provided
WorkerInfo or name. When using RPC with DDP, sometimes it might just be easier
to refer to everything in terms of ranks since DDP doesn't support names yet.

As a result, support a `to` parameter in the RPC APIs which allow for
specifying a rank as well would be helpful.

Differential Revision: [D24264989](https://our.internmc.facebook.com/intern/diff/D24264989/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D24264989/)!

ghstack-source-id: 114127380
Pull Request resolved: https://github.com/pytorch/pytorch/pull/46221